[chore] prometheusreceiver: remove metrics builder, incorporate into transaction
